rss_mail_summarizer/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
babel==2.16.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
courlan==1.3.2
dateparser==1.2.0
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain-core==0.3.31
langchain-google-genai==2.0.9
langsmith==0.3.1
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
orjson==3.10.15
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tenacity==9.0.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
zstandard==0.23.0
